+++ CHANGELOG.md
@@ -2,6 +2,7 @@ walkers 5.8.0
 ================
 - add AlexsCaves mobs to flying.json
 - add a config option to prevent hostile mobs from attacking neutral
+- fix showPlayerNametag with Team settings
 
 walkers 5.7
 ================